Split Transactions Controller unit tests
